Update: GradeEntity and StudentEntity classes

+++ src/System/Model/Academy/GradeEntity.cs
@@ -5,7 +5,7 @@ public class GradeEntity
     public int gradeId { get; set; }
     public string? studentId { get; set; }
     public string? subjectId { get; set; }
-    public int grade { get; set; }
-    public int conductGrade { get; set; }
+    public decimal grade { get; set; }
+    public decimal conductGrade { get; set; }
     public string label { get; set; } = null!;
 }
